function without aurgument with return value
In this example, the calculate_sum() function does not take any arguments. It performs a simple addition operation on two numbers (num1 and num2) and stores the result in the sum variable. Finally, it returns the value of sum. The function is then called and the returned value is stored in the result variable, which is printed to the console.